+++ packages/eslint-config/CHANGELOG.md
@@ -1,8 +1,29 @@
 # @fuelrats/eslint-config Changelog
 
 
-## Unreleased
+### Unreleased
 
+#### Additions
+* Warn when a lint rule is being disabled that doesn't need to be.
+* Enable ESLint v7 rules
+  * `default-case-last`
+  * `no-useless-backreference`
+  * `no-promise-executor-return`
+  * `no-unreachable-loop`
+* Recognize the path alias `~` as an `internal` import in import ordering.
+
+
+#### Changes
+* Better and more consistent restricted syntax messages.
+* Disallow explicit null checks via `no-restricted syntax` instead of abusing the combination of `eqeqeq`'s null option and `no-eq-null`
+* add `inheritdoc` to certain `jsdoc` rule `exemptedBy` options so it's consistent with the default config
+* make `import/order` case insensitive sort.
+* Remove all node-specific rules since they are now deprecated.
+* Improved settings for import plugin.
+* Use `id-denylist` over `id-blacklist`
+  * Even though this rule is off we'd rather not have deprecated rule names :D
+* Ensure our config is JSON serializable by replacing `Infinity` with `Number.MAX_SAFE_INTEGER`
+* Enable `ignoreGlobals` for `camelcase` rule
 
 
 ### 2.1.0